Move trait implementation in right place
