+++ crates/fuel-core/Cargo.toml
@@ -67,7 +67,6 @@ fuel-core-storage = { path = "./../storage", features = ["test-helpers"] }
 fuel-core-trace = { path = "./../trace" }
 fuel-core-types = { path = "./../types", features = ["test-helpers"] }
 mockall = { workspace = true }
-primitive-types = { workspace = true, default-features = false }
 proptest = { workspace = true }
 test-case = { workspace = true }
 test-strategy = { workspace = true }
