+++ tests/tests/lib.rs
@@ -26,5 +26,6 @@ mod trigger_integration;
 mod tx;
 #[cfg(feature = "p2p")]
 mod tx_gossip;
+mod vm_storage;
 
 fuel_core_trace::enable_tracing!();
